[rename] rename auth
