+++ src/main/java/com/nexters/buyornot/module/auth/api/AuthController.java
@@ -3,8 +3,8 @@ package com.nexters.buyornot.module.auth.api;
 import com.nexters.buyornot.global.common.codes.SuccessCode;
 import com.nexters.buyornot.global.common.response.ApiResponse;
 import com.nexters.buyornot.module.auth.application.OAuthLoginService;
-import com.nexters.buyornot.module.auth.infra.KakaoLoginParams;
-import com.nexters.buyornot.module.auth.model.AuthTokens;
+import com.nexters.buyornot.module.auth.api.dto.request.KakaoLoginParams;
+import com.nexters.buyornot.module.auth.api.dto.response.AuthTokens;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
