+++ src/main/java/com/nexters/buyornot/module/auth/api/dto/response/AuthTokens.java
@@ -1,4 +1,4 @@
-package com.nexters.buyornot.module.auth.model;
+package com.nexters.buyornot.module.auth.api.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
